requirements.txt
psycopg2-binary
Flask
Flask-SQLAlchemy
flask-migrate
flask-caching
python-dotenv

